+++ src/main/java/kg/attractor/jobsearch/controller/VacancyController.java
@@ -33,7 +33,7 @@ public class VacancyController {
     }
 
     @GetMapping("vacancies/{vacancyId}")
-    public String getInfo(@PathVariable long vacancyId, Model model, Authentication authentication) {
+    public String getInfo(@PathVariable Integer vacancyId, Model model, Authentication authentication) {
         MvcConrollersUtil.authCheckAndAddAttributes(
                 model,
                 authentication,
